// SkilTracker/src/app/models/associate-dashboard.ts
export class AssociateDashboard {
    associateid: number;
    associatename: string;
    email: string;
    mobile: string;
    image: string;
    statusgreen: number;
    statusblue: number;
    statusred: number;
    level1: number;
    level2: number;
    level3: number;
    remarks: string;
    strength: string;
    weakness: string;
    skillset: string;
}